Factor out irydium components into a seperate components section
Lumping everything into examples is a little confusing.
