+++ packages/site/static/components/plotlyjs.md
