+++ packages/site/static/components/vega-embed.md
