--- packages/site/static/examples
@@ -1,1 +1,0 @@
-../src/examples
