+++ packages/site/static/examples/document-pyodide.irmd
