+++ packages/site/static/examples/pyodide.md
